add test and stuff

+++ app/Models/Ingredient.php
@@ -4,13 +4,14 @@ namespace App\Models;
 
 use App\Models\Pizza;
 use Helper;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use Storage;
 
 class Ingredient extends Model
 {
-    use SoftDeletes;
+    use HasFactory, SoftDeletes;
 
     protected $guarded = ['id', 'created_at', 'updated_at', 'deleted_at'];
 
@@ -35,7 +36,7 @@ class Ingredient extends Model
         parent::boot();
 
         static::deleting(function ($ingredient) {
-            Helper::deleteFile($ingredient->image);
+            if ($ingredient->image) Helper::deleteFile($ingredient->image);
         });
     }
 }
